moving 3d view stuff into 3dview dir, extracting shape field names, instrument name etc
