+++ nexus_constructor/3dview/instrument_zooming_3d_window.py
