--- nexus_constructor/file_dialog_options.py
@@ -1,3 +1,0 @@
-from PySide2.QtWidgets import QFileDialog
-
-FILE_DIALOG_NATIVE = QFileDialog.DontUseNativeDialog
