+++ nexus_constructor/model/instrument.py
@@ -1,6 +1,5 @@
-from nexus_constructor.common_attrs import CommonKeys
+from nexus_constructor.common_attrs import CommonKeys, INSTRUMENT_NAME
 from nexus_constructor.model.component import Component
-from nexus_constructor.model.entry import INSTRUMENT_NAME
 from nexus_constructor.model.group import Group
 from typing import Dict, Any
