+++ nexus_constructor/model/transformation.py
@@ -5,10 +5,14 @@ import numpy as np
 from PySide2.Qt3DCore import Qt3DCore
 from PySide2.QtGui import QVector3D, QMatrix4x4
 
-from nexus_constructor.common_attrs import CommonAttrs, CommonKeys, NodeType
+from nexus_constructor.common_attrs import (
+    CommonAttrs,
+    CommonKeys,
+    NodeType,
+    TransformationType,
+)
 from nexus_constructor.model.dataset import Dataset
 from nexus_constructor.model.value_type import ValueTypes
-from nexus_constructor.transformation_types import TransformationType
 
 from typing import TYPE_CHECKING
