--- nexus_constructor/transformation_types.py
@@ -1,3 +1,0 @@
-class TransformationType:
-    TRANSLATION = "Translation"
-    ROTATION = "Rotation"
